Improve CSRF token grabbing
Previously a regex was used to grab the CSRF info. Since it depended on
not only the markup, but how the markup was formatted, it was brittle.

Since `BeautifulSoup` is already being used to parse bike components, it
makes sense to also use it to grab the CSRF info. The built-in
'html.parser' library chokes on non-XHTML meta tags ("<meta>" vs.
"<meta/>") so the `html5lib` parser is used instead.

As a sidenote, `lxml` would also work and be faster, but requires some
system libraries that make setup a little less friendly. Since parsing
speed shouldn't be much of an issue and the pages are relatively small,
this seems like an acceptable tradeoff.