Allow WebClient to be recreated without logging in again (#9)
As described in #2, logging into the Strava web interface creates 3 cookies:

- `strava_remember_token` (required) is a JWT.
- `strava_remember_id` (required) is the logged in user's id. It can be
  recreated from the "sub" field in the `strava_remember_token`.
- `_strava4_session` (not required) is just a session token. It will
  automatcally be created by the server if not provided.

We can use our understanding of the Strava cookies and the CSRF token to make
it possible to easily recreate the `WebClient` without repeating the login
process or retaining the user's password.

Among other things, caching the JWT and reusing it can avoid
rate-limiting/lockout, which will happen if you log into the Strava web
interface too frequently.

Co-authored-by: Carey Metcalfe <carey@cmetcalfe.ca>