Revert to using html.parser instead of html5lib
The original reason for using `html5lib` (added in d9cf8cda) was that
the built-in `html.parser` had issues parsing `<meta ...>` tags without the
proper trailing slash. This caused problems when attempting to parse the
CSRF token out of the login page.

The login page has been updated since then to include proper `<meta ... />`
tags so we can go back to using the built-in `html.parser` and drop the
`html5lib` dependency.

Fixes #10